rework the taint logic (#3)
* rework the taint logic

There was previous incorrect assumptions about multiple taints
with the same key being valid. That was incorrect, and would not work
for cases with multiple daemonsets that need to be waited on.

This changes the structure so there is one big taint that all daemonsets
must each become ready for. Once all daemonset pods are ready, that node
then gets the taint removed, and regular scheduling can occur.

This is slightly less flexible than the previous design, but it fixes
the previous assumptions for 2+ daemonsets cases.